requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.2
Django==4.2.2
django-cors-headers==4.1.0
django-cron==0.6.0
django-environ==0.10.0
django-model-utils==4.3.1
django-push-notifications==3.0.0
django-ratelimit==4.0.0
django-ratelimit2==0.4.2
django-redis==5.3.0
django-snitch==2.3.0
django-timezone-field==5.1
django-websocket-notifications==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
generics==6.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
msgpack==1.0.5
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
zope.interface==6.0
